fixing user profile completed activities

+++ back-end/models/activity.js
@@ -3,8 +3,8 @@ module.exports = function(sequelize, DataTypes) {
   var Activity = sequelize.define('Activity', {
     name: DataTypes.STRING,
     description: DataTypes.TEXT,
-    longitude: DataTypes.INTEGER,
-    latitude: DataTypes.INTEGER
+    longitude: DataTypes.FLOAT,
+    latitude: DataTypes.FLOAT
   }, {
     classMethods: {
       associate: function(models) {
